+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 8.0.5 (TBD)
 
+* DAO column range support (including geometry envelopes & bounding boxes) to build where clauses & args for queries
 * ogc-api-features-json-ios version 4.2.5
 
 ## [8.0.4](https://github.com/ngageoint/geopackage-ios/releases/tag/8.0.4) (11-15-2023)